requirements.txt
netCDF4==1.6.5
numpy==1.23.3
google-cloud-compute==1.15.0
google-cloud-storage==2.14.0
ruff==0.1.13
pre-commit==3.5.0
rich==13.7.1
cftime==1.6.3
matplotlib==3.8.4
xarray==2024.3.0
